Added rounded corners option for progress bar (#1119)
* Added rounded corners option for progress bar

* Removed unnecessary spaces. Removed corner radius > 0 checks. Fixed sanity check.

* Fix transposed LOG_W arguments. Remove unneeded parentheses. Add TODOs with information regarding potential blurry lines.

+++ src/draw.h
@@ -12,7 +12,7 @@ void draw_setup(void);
 
 void draw(void);
 
-void draw_rounded_rect(cairo_t *c, int x, int y, int width, int height, int corner_radius, double scale, bool first, bool last);
+void draw_rounded_rect(cairo_t *c, float x, float y, int width, int height, int corner_radius, double scale, bool first, bool last);
 
 // TODO get rid of this function by passing scale to everything that needs it.
 double draw_get_scale(void);
